Move the test folder to the top level
