+++ tests/header.test.tsx
@@ -5,7 +5,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import Header from '../components/header/Header';
+import Header from '../src/components/header/Header';
 import { MemoryRouter } from 'react-router-dom';
 
 describe('Header component', () => {
